+++ src/main/java/com/example/whostolemyfood/store/application/service/StoreServiceV1.java
@@ -1,5 +1,7 @@
 package com.example.whostolemyfood.store.application.service;
 
+import com.example.whostolemyfood.global.exception.CustomException;
+import com.example.whostolemyfood.global.exception.ErrorCode;
 import com.example.whostolemyfood.store.domain.entity.StoreEntity;
 import com.example.whostolemyfood.store.domain.entity.StoreStatus;
 import com.example.whostolemyfood.store.domain.repository.StoreRepository;
@@ -8,6 +10,10 @@ import com.example.whostolemyfood.store.presentation.dto.request.ReqUpdateStoreD
 import com.example.whostolemyfood.store.presentation.dto.response.ResCreateStoreDtoV1;
 import com.example.whostolemyfood.store.presentation.dto.response.ResGetStoreDtoV1;
 import com.example.whostolemyfood.store.presentation.dto.response.ResGetStoreListDtoV1;
+import com.example.whostolemyfood.user.application.security.AuthUser;
+import com.example.whostolemyfood.user.domain.entity.UserEntity;
+import com.example.whostolemyfood.user.domain.entity.UserRole;
+import com.example.whostolemyfood.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,19 +27,26 @@ import java.util.UUID;
 public class StoreServiceV1 {
 
     private final StoreRepository storeRepository;
+    private final UserRepository userRepository;
 
     // Owner Only
     // 스토어 생성
     @Transactional
-    public ResCreateStoreDtoV1 createStore(ReqCreateStoreDtoV1 request) {
+    public ResCreateStoreDtoV1 createStore(ReqCreateStoreDtoV1 request, AuthUser authUser) {
         // 유저 권한 확인 로직
+        if (authUser.role() != UserRole.OWNER) {
+            throw new CustomException(ErrorCode.ACCESS_DENIED);
+        }
+        UserEntity owner = userRepository.findById(authUser.userId())
+                .orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
 
         // 스토어 이름 중복 확인
         if (storeRepository.existsByNameAndIsDeletedFalse(request.getName())) {
-            throw new IllegalArgumentException("이미 존재하는 스토어 이름입니다 : " + request.getName());
+            throw new CustomException(ErrorCode.STORE_DUPLICATION_NAME);
         }
 
         StoreEntity store = StoreEntity.builder()
+                .user(owner)
                 .name(request.getName())
                 .address(request.getAddress())
                 .phone(request.getPhone())
@@ -53,11 +66,12 @@ public class StoreServiceV1 {
     @Transactional(readOnly = true)
     public ResGetStoreDtoV1 getStore(UUID storeId) {
         StoreEntity store = storeRepository.findByStoreIdAndIsHiddenFalseAndIsDeletedFalse(storeId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 스토어를 찾을 수 없습니다"));
+                .orElseThrow(() -> new CustomException(ErrorCode.STORE_NOT_FOUND));
 
         return ResGetStoreDtoV1.from(store);
     }
 
+    // All
     // 스토어 목록 조회
     @Transactional(readOnly = true)
     public Page<ResGetStoreListDtoV1> getStores(Pageable pageable) {
@@ -66,41 +80,58 @@ public class StoreServiceV1 {
         return stores.map(ResGetStoreListDtoV1::from);
     }
 
-    // owner, manager, master
+    // Owner, Manager, Master
     // 스토어 수정
     @Transactional
-    public ResGetStoreDtoV1 updateStore(UUID storeId, ReqUpdateStoreDtoV1 request) {
+    public ResGetStoreDtoV1 updateStore(UUID storeId, ReqUpdateStoreDtoV1 request, AuthUser authUser) {
         StoreEntity store = storeRepository.findByStoreIdAndIsDeletedFalse(storeId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 스토어를 찾을 수 없습니다"));
-        // 유저 권한 확인
+                .orElseThrow(() -> new CustomException(ErrorCode.STORE_NOT_FOUND));
+        // 권한 확인
+        validateStoreAccess(store, authUser);
 
         // 이름 수정시에 발동
         if (!store.getName().equals(request.getName())) {
-            if (storeRepository.existsByNameAndIsDeletedFalse(store.getName())) {
-                throw new IllegalArgumentException("이미 존재하는 스토어 이름입니다 : " + request.getName());
+            if (storeRepository.existsByNameAndIsDeletedFalse(request.getName())) {
+                throw new CustomException(ErrorCode.STORE_DUPLICATION_NAME);
             }
         }
 
-        store.updateAllFields(request);
+        store.updateStore(request);
 
         return ResGetStoreDtoV1.from(store);
     }
 
+    // Owner, Manager, Master
+    // 스토어 숨김 / 노출
+    @Transactional
+    public void hiddenStore(UUID storeId, AuthUser authUser) {
+        StoreEntity store = storeRepository.findByStoreIdAndIsDeletedFalse(storeId)
+                .orElseThrow(()-> new CustomException(ErrorCode.STORE_NOT_FOUND));
+        // 권한 확인
+        validateStoreAccess(store, authUser);
+
+        store.toggleIsHidden();
+    }
+
+    // Owner, Manager, Master
     // 스토어 삭제
     @Transactional
-    public void deleteStore(UUID storeId) {
-        // 유저 권한 확인
+    public void deleteStore(UUID storeId,AuthUser authUser) {
         StoreEntity store = storeRepository.findByStoreIdAndIsDeletedFalse(storeId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 스토어가 존재하지 않습니다"));
+                .orElseThrow(() -> new CustomException(ErrorCode.STORE_NOT_FOUND));
+        // 권한 확인
+        validateStoreAccess(store, authUser);
+
         store.deleteByOwnerAndMaster(storeId);
     }
 
-    // 스토어 숨김 / 노출
-    @Transactional
-    public void hiddenStore(UUID storeId) {
-        StoreEntity store = storeRepository.findByStoreIdAndIsDeletedFalse(storeId)
-                .orElseThrow(()-> new IllegalArgumentException("해당 스토어가 존재하지 않습니다"));
-
-        store.toggleIsHidden();
+    // 권한 확인
+    private void validateStoreAccess(StoreEntity store, AuthUser authUser) {
+        if (authUser.role() == UserRole.CUSTOMER) {
+            throw new CustomException(ErrorCode.ACCESS_DENIED);
+        }
+        if (authUser.role() == UserRole.OWNER && !store.getUser().getId().equals(authUser.userId())) {
+            throw new CustomException(ErrorCode.STORE_NOT_OWNER);
+        }
     }
 }
